pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "realtime-kafka-pipeline"
version = "0.1.0"
dependencies = [
    "pyodbc",
    "python-dotenv",
    "pyyaml",
    "faker",
    "confluent-kafka",
    "fastavro"
]

[tool.setuptools.packages.find]
include = ["kafka_producer", "sqlScripts"]
